Segrega os endpoints 'criarConta' e 'obterConta'
